split `cpython` portion of `setobject.h` off
