+++ pyo3-ffi/src/cpython/mod.rs
@@ -43,6 +43,7 @@ pub(crate) mod pythonrun;
 pub(crate) mod floatobject;
 pub(crate) mod pyframe;
 pub(crate) mod pyhash;
+pub(crate) mod setobject;
 pub(crate) mod traceback;
 pub(crate) mod tupleobject;
 pub(crate) mod unicodeobject;
@@ -90,6 +91,7 @@ pub use self::pylifecycle::*;
 pub use self::pymem::*;
 pub use self::pystate::*;
 pub use self::pythonrun::*;
+pub use self::setobject::*;
 pub use self::traceback::*;
 pub use self::tupleobject::*;
 pub use self::unicodeobject::*;
