+++ pyo3-ffi/src/setobject.rs
@@ -1,63 +1,6 @@
-#[cfg(Py_GIL_DISABLED)]
-use crate::cpython::pyatomic::_Py_atomic_load_ssize_relaxed;
 use crate::object::*;
-#[cfg(not(any(Py_LIMITED_API, PyPy, GraalPy)))]
-use crate::pyport::Py_hash_t;
 use crate::pyport::Py_ssize_t;
 use core::ffi::c_int;
-
-pub const PySet_MINSIZE: usize = 8;
-
-#[cfg(not(any(Py_LIMITED_API, PyPy, GraalPy)))]
-#[repr(C)]
-#[derive(Debug)]
-pub struct setentry {
-    pub key: *mut PyObject,
-    pub hash: Py_hash_t,
-}
-
-#[cfg(not(any(Py_LIMITED_API, PyPy, GraalPy)))]
-#[repr(C)]
-#[derive(Debug)]
-pub struct PySetObject {
-    pub ob_base: PyObject,
-    pub fill: Py_ssize_t,
-    pub used: Py_ssize_t,
-    pub mask: Py_ssize_t,
-    pub table: *mut setentry,
-    pub hash: Py_hash_t,
-    pub finger: Py_ssize_t,
-    pub smalltable: [setentry; PySet_MINSIZE],
-    pub weakreflist: *mut PyObject,
-}
-
-#[inline]
-#[cfg(all(not(any(PyPy, GraalPy)), not(Py_LIMITED_API)))]
-pub(crate) unsafe fn _PySet_CAST(so: *mut PyObject) -> *mut PySetObject {
-    debug_assert_eq!(PyAnySet_Check(so), 1);
-    so.cast()
-}
-
-#[inline]
-#[cfg(all(not(any(PyPy, GraalPy)), not(Py_LIMITED_API)))]
-pub unsafe fn PySet_GET_SIZE(so: *mut PyObject) -> Py_ssize_t {
-    let so = _PySet_CAST(so);
-    #[cfg(Py_GIL_DISABLED)]
-    {
-        _Py_atomic_load_ssize_relaxed(&raw const (*so).used)
-    }
-    #[cfg(not(Py_GIL_DISABLED))]
-    {
-        (*so).used
-    }
-}
-
-// skipped _PySet_Dummy
-
-extern_libpython! {
-    // skipped non-limited _PySet_NextEntry
-    // skipped non-limited _PySet_Update
-}
 
 #[cfg(not(RustPython))]
 extern_libpython! {
